refactor(web): step 11b — extract useComposerLayout (reflow, radius clock, height morph) from chat-pane
